Feature/order create

+++ src/main/java/team7/delivery/repository/MenuRepository.java
@@ -4,5 +4,5 @@ package team7.delivery.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import team7.delivery.entity.Menu;
 
-public interface MenuRepository extends JpaRepository<Menu, Integer> {
+public interface MenuRepository extends JpaRepository<Menu, Long> {
 }
